[util] Move assertEl to a new utility module
